+++ berserk/clients/opening_explorer.py
@@ -1,104 +1,20 @@
 from __future__ import annotations
 
-from typing import Iterator, cast, List, Literal, TypedDict
+from typing import Iterator, cast, Literal
 import requests
 import logging
 
 from .base import BaseClient
+from ..types import (
+    OpeningStatistic,
+    OpeningExplorerVariant,
+    Speed,
+    OpeningExplorerRating,
+)
 
 logger = logging.getLogger("berserk.client.opening_explorer")
 
 EXPLORER_URL = "https://explorer.lichess.ovh"
-
-OpeningExplorerVariant = Literal[
-    "standard",
-    "chess960",
-    "crazyhouse",
-    "antichess",
-    "atomic",
-    "horde",
-    "kingOfTheHill",
-    "racingKings",
-    "threeCheck",
-    "fromPosition",
-]
-
-Speed = Literal[
-    "ultraBullet", "bullet", "blitz", "rapid", "classical", "correspondence"
-]
-
-OpeningExplorerRating = Literal[
-    "0", "1000", "1200", "1400", "1600", "1800", "2000", "2200", "2500"
-]
-
-
-class Opening(TypedDict):
-    # The eco code of this opening
-    eco: str
-    # The name of this opening
-    name: str
-
-
-class Player(TypedDict):
-    # The name of the player
-    name: str
-    # The rating of the player during the game
-    rating: int
-
-
-class Game(TypedDict):
-    # The move in Universal Chess Interface notation
-    uci: str
-    # The id of the game
-    id: str
-    # The winner of the game. Draw if None
-    winner: Literal["white", "black"] | None
-    # The speed of the game
-    speed: Speed
-    # The type of game
-    mode: Literal["rated", "casual"]
-    # The black player
-    black: Player
-    # The white player
-    white: Player
-    # The year of the game
-    year: int
-    # The month and year of the game. For example "2023-06"
-    month: str
-
-
-class Move(TypedDict):
-    # The move in Universal Chess Interface notation
-    uci: str
-    # The move in algebraic notation
-    san: str
-    # The average rating of games in the position after this move
-    averageRating: str
-    # The number of white winners after this move
-    white: int
-    # The number of black winners after this move
-    black: int
-    # The number of draws after this move
-    draws: int
-    # The game where the move was played
-    game: Game | None
-
-
-class OpeningStatistic(TypedDict):
-    # Number of game won by white from this position
-    white: int
-    # Number of game won by black from this position
-    draws: int
-    # Number draws from this position
-    black: int
-    # Opening info of this position
-    opening: Opening | None
-    # The list of moves played by players from this position
-    moves: List[Move]
-    # recent games with this opening
-    recentGames: List[Game]
-    # top rating games with this opening
-    topGames: List[Game]
 
 
 class OpeningExplorer(BaseClient):
@@ -111,9 +27,9 @@ class OpeningExplorer(BaseClient):
         self,
         variant: OpeningExplorerVariant = "standard",
         position: str | None = None,
-        play: List[str] | None = None,
-        speeds: List[Speed] | None = None,
-        ratings: List[OpeningExplorerRating] | None = None,
+        play: list[str] | None = None,
+        speeds: list[Speed] | None = None,
+        ratings: list[OpeningExplorerRating] | None = None,
         since: str | None = None,
         until: str | None = None,
         moves: int | None = None,
@@ -155,7 +71,7 @@ class OpeningExplorer(BaseClient):
     def get_masters_games(
         self,
         position: str | None = None,
-        play: List[str] | None = None,
+        play: list[str] | None = None,
         since: int | None = None,
         until: int | None = None,
         moves: int | None = None,
@@ -181,9 +97,9 @@ class OpeningExplorer(BaseClient):
         color: Literal["white", "black"],
         variant: OpeningExplorerVariant | None = None,
         position: str | None = None,
-        play: List[str] | None = None,
-        speeds: List[Speed] | None = None,
-        ratings: List[OpeningExplorerRating] | None = None,
+        play: list[str] | None = None,
+        speeds: list[Speed] | None = None,
+        ratings: list[OpeningExplorerRating] | None = None,
         since: int | None = None,
         until: int | None = None,
         moves: int | None = None,
@@ -229,9 +145,9 @@ class OpeningExplorer(BaseClient):
         color: Literal["white", "black"],
         variant: OpeningExplorerVariant | None = None,
         position: str | None = None,
-        play: List[str] | None = None,
-        speeds: List[Speed] | None = None,
-        ratings: List[OpeningExplorerRating] | None = None,
+        play: list[str] | None = None,
+        speeds: list[Speed] | None = None,
+        ratings: list[OpeningExplorerRating] | None = None,
         since: int | None = None,
         until: int | None = None,
         moves: int | None = None,